# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Roleplay Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AdminSettings {
  id: number;
  aiProviderName: string;
  aiProviderEndpoint: string;
  /** @nullable */
  aiProviderApiKey?: string | null;
  maxContextSize: number;
  contextSummarizationThreshold: number;
  updatedAt: string;
}

export interface UpdateAdminSettingsBody {
  /** @nullable */
  aiProviderName?: string | null;
  /** @nullable */
  aiProviderEndpoint?: string | null;
  /** @nullable */
  aiProviderApiKey?: string | null;
  /** @nullable */
  maxContextSize?: number | null;
  /** @nullable */
  contextSummarizationThreshold?: number | null;
}

export interface AiModel {
  id: number;
  modelId: string;
  displayName: string;
  enabled: boolean;
  isDefault: boolean;
  createdAt: string;
}

export interface CreateAiModelBody {
  modelId: string;
  displayName: string;
  isDefault?: boolean;
}

export interface UpdateAiModelBody {
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  enabled?: boolean | null;
  /** @nullable */
  isDefault?: boolean | null;
}

export interface AdminStats {
  totalUsers: number;
  totalCharacters: number;
  totalChats: number;
  totalMessages: number;
  activeModels: number;
}

export interface Character {
  id: number;
  ownerUserId: string;
  name: string;
  age: number;
  personality: string;
  description: string;
  /** @nullable */
  profilePicture?: string | null;
  backgroundStory: string;
  lore: string;
  introMessage: string;
  tags: string[];
  chatCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateCharacterBody {
  name: string;
  age: number;
  personality: string;
  description: string;
  /** @nullable */
  profilePicture?: string | null;
  backgroundStory: string;
  lore: string;
  introMessage: string;
  tags?: string[];
}

export interface UpdateCharacterBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  personality?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  profilePicture?: string | null;
  /** @nullable */
  backgroundStory?: string | null;
  /** @nullable */
  lore?: string | null;
  /** @nullable */
  introMessage?: string | null;
  /** @nullable */
  tags?: string[] | null;
}

export interface Persona {
  id: number;
  ownerUserId: string;
  name: string;
  age: number;
  personality: string;
  description: string;
  lore: string;
  tags: string[];
  isMain: boolean;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export interface CreatePersonaBody {
  name: string;
  age: number;
  personality: string;
  description: string;
  lore: string;
  tags?: string[];
  /** @nullable */
  avatarUrl?: string | null;
}

export interface UpdatePersonaBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  personality?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  lore?: string | null;
  /** @nullable */
  tags?: string[] | null;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface Tag {
  id: number;
  name: string;
  usageCount: number;
}

export interface CreateTagBody {
  name: string;
}

export interface Chat {
  id: number;
  userId: string;
  characterId: number;
  personaId: number;
  character?: Character;
  persona?: Persona;
  /** @nullable */
  systemPrompt?: string | null;
  /** @nullable */
  modelId?: number | null;
  /** @nullable */
  lastMessageAt?: string | null;
  createdAt: string;
}

export interface CreateChatBody {
  characterId: number;
  personaId: number;
  /** @nullable */
  modelId?: number | null;
  /** @nullable */
  systemPrompt?: string | null;
}

export interface UpdateChatBody {
  /** @nullable */
  systemPrompt?: string | null;
  /** @nullable */
  modelId?: number | null;
}

export interface Message {
  id: number;
  chatId: number;
  role: string;
  content: string;
  /** @nullable */
  modelUsed?: string | null;
  createdAt: string;
}

export interface ChatWithMessages {
  id: number;
  userId: string;
  characterId: number;
  personaId: number;
  character?: Character;
  persona?: Persona;
  /** @nullable */
  systemPrompt?: string | null;
  /** @nullable */
  modelId?: number | null;
  messages: Message[];
  /** @nullable */
  lastMessageAt?: string | null;
  createdAt: string;
}

export interface SendMessageBody {
  content: string;
  /** @nullable */
  modelId?: number | null;
}

export interface MessagePair {
  userMessage: Message;
  assistantMessage: Message;
  modelUsed: string;
  blocked: boolean;
  /** @nullable */
  blockReason?: string | null;
}

export interface GroupMember {
  id: number;
  groupId: number;
  memberType: string;
  /** @nullable */
  characterId?: number | null;
  /** @nullable */
  personaId?: number | null;
  /** @nullable */
  userId?: string | null;
  character?: Character;
  persona?: Persona;
  joinedAt: string;
}

export interface GroupChat {
  id: number;
  name: string;
  ownerUserId: string;
  members: GroupMember[];
  createdAt: string;
}

export interface GroupMessage {
  id: number;
  groupId: number;
  senderType: string;
  /** @nullable */
  senderId?: number | null;
  /** @nullable */
  senderUserId?: string | null;
  senderName: string;
  content: string;
  /** @nullable */
  modelUsed?: string | null;
  createdAt: string;
}

export interface GroupChatWithMessages {
  id: number;
  name: string;
  ownerUserId: string;
  members: GroupMember[];
  messages: GroupMessage[];
  createdAt: string;
}

export interface CreateGroupBody {
  name: string;
  personaId: number;
  characterIds?: number[];
}

export interface AddGroupMemberBody {
  memberType: string;
  /** @nullable */
  characterId?: number | null;
  /** @nullable */
  personaId?: number | null;
}

export interface SendGroupMessageBody {
  content: string;
  personaId: number;
}

export interface GroupInvite {
  id: number;
  groupId: number;
  inviterPersonaId: number;
  inviteePersonaId: number;
  status: string;
  group?: GroupChat;
  inviterPersona?: Persona;
  inviteePersona?: Persona;
  createdAt: string;
}

export interface SendGroupInviteBody {
  inviterPersonaId: number;
  inviteePersonaId: number;
}

export interface RespondGroupInviteBody {
  accept: boolean;
}

export interface UploadResult {
  url: string;
  filename: string;
}

export interface DmConversation {
  personaId: number;
  persona?: Persona;
  /** @nullable */
  lastMessage?: string | null;
  /** @nullable */
  lastMessageAt?: string | null;
  unreadCount: number;
}

export interface SendDmBody {
  toPersonaId: number;
  fromPersonaId: number;
  content: string;
}

export interface DmMessage {
  id: number;
  fromPersonaId: number;
  toPersonaId: number;
  fromPersona?: Persona;
  content: string;
  createdAt: string;
}

export type ListCharactersParams = {
  /**
   * @nullable
   */
  search?: string | null;
  /**
   * @nullable
   */
  tags?: string | null;
  /**
   * @nullable
   */
  mine?: string | null;
};

export type SearchPersonasParams = {
  /**
   * @nullable
   */
  q?: string | null;
};

export type UploadImageBody = {
  file?: Blob;
};