# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Roleplay Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/settings:
    get:
      operationId: getAdminSettings
      tags: [admin]
      summary: Get admin settings
      responses:
        "200":
          description: Admin settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettings"
    patch:
      operationId: updateAdminSettings
      tags: [admin]
      summary: Update admin settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAdminSettingsBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettings"
  /admin/models:
    get:
      operationId: listAdminModels
      tags: [admin]
      summary: List all AI models
      responses:
        "200":
          description: Models list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiModel"
    post:
      operationId: createAdminModel
      tags: [admin]
      summary: Add a new AI model
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAiModelBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiModel"
  /admin/models/{id}:
    patch:
      operationId: updateAdminModel
      tags: [admin]
      summary: Update an AI model (rename, enable/disable)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAiModelBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiModel"
    delete:
      operationId: deleteAdminModel
      tags: [admin]
      summary: Delete an AI model
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /characters:
    get:
      operationId: listCharacters
      tags: [characters]
      summary: List/search characters
      parameters:
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: tags
          in: query
          schema:
            type: ["string", "null"]
        - name: mine
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Characters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Character"
    post:
      operationId: createCharacter
      tags: [characters]
      summary: Create a character
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCharacterBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
...
  /characters/{id}:
    get:
      operationId: getCharacter
      tags: [characters]
      summary: Get a character
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
    patch:
      operationId: updateCharacter
      tags: [characters]
      summary: Update a character (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCharacterBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
    delete:
      operationId: deleteCharacter
      tags: [characters]
      summary: Delete a character (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Personas ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    AdminSettings:
      type: object
      properties:
        id:
          type: integer
        aiProviderName:
          type: string
        aiProviderEndpoint:
          type: string
        aiProviderApiKey:
          type: ["string", "null"]
        maxContextSize:
          type: integer
        contextSummarizationThreshold:
          type: integer
        updatedAt:
          type: string
      required: [id, aiProviderName, aiProviderEndpoint, maxContextSize, contextSummarizationThreshold, updatedAt]
    UpdateAdminSettingsBody:
      type: object
      properties:
        aiProviderName:
          type: ["string", "null"]
        aiProviderEndpoint:
          type: ["string", "null"]
        aiProviderApiKey:
          type: ["string", "null"]
        maxContextSize:
          type: ["integer", "null"]
        contextSummarizationThreshold:
          type: ["integer", "null"]
    AiModel:
      type: object
      properties:
        id:
          type: integer
        modelId:
          type: string
        displayName:
          type: string
        enabled:
          type: boolean
        isDefault:
          type: boolean
        createdAt:
          type: string
      required: [id, modelId, displayName, enabled, isDefault, createdAt]
    CreateAiModelBody:
      type: object
      properties:
        modelId:
          type: string
        displayName:
          type: string
        isDefault:
          type: boolean
      required: [modelId, displayName]
    UpdateAiModelBody:
      type: object
      properties:
        displayName:
          type: ["string", "null"]
        enabled:
          type: ["boolean", "null"]
        isDefault:
          type: ["boolean", "null"]
...
    Character:
      type: object
      properties:
        id:
          type: integer
        ownerUserId:
          type: string
        name:
          type: string
        age:
          type: integer
        personality:
          type: string
        description:
          type: string
        profilePicture:
          type: ["string", "null"]
        backgroundStory:
          type: string
        lore:
          type: string
        introMessage:
          type: string
        tags:
          type: array
          items:
            type: string
        chatCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, ownerUserId, name, age, personality, description, backgroundStory, lore, introMessage, tags, chatCount, createdAt, updatedAt]
    CreateCharacterBody:
      type: object
      properties:
        name:
          type: string
        age:
          type: integer
        personality:
          type: string
        description:
          type: string
        profilePicture:
          type: ["string", "null"]
        backgroundStory:
          type: string
        lore:
          type: string
        introMessage:
          type: string
        tags:
          type: array
          items:
            type: string
      required: [name, age, personality, description, backgroundStory, lore, introMessage]
    UpdateCharacterBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        age:
          type: ["integer", "null"]
        personality:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        profilePicture:
          type: ["string", "null"]
        backgroundStory:
          type: ["string", "null"]
        lore:
          type: ["string", "null"]
        introMessage:
          type: ["string", "null"]
        tags:
          type: ["array", "null"]
          items:
            type: string
...