# lib/api-zod/src/generated/types/adminSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Roleplay Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminSettings {
  id: number;
  aiProviderName: string;
  aiProviderEndpoint: string;
  /** @nullable */
  aiProviderApiKey?: string | null;
  maxContextSize: number;
  contextSummarizationThreshold: number;
  updatedAt: string;
}