# lib/api-zod/src/generated/types/updateAiModelBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Roleplay Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateAiModelBody {
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  enabled?: boolean | null;
  /** @nullable */
  isDefault?: boolean | null;
}